S3 - add documentation #103
