+++ s3/src/main/resources/reference.conf
@@ -5,14 +5,14 @@ akka.stream.alpakka.s3 {
   # location for temporary files, if buffer is set to "disk". If empty, uses the standard java temp path.
   disk-buffer-path = ""
 
-  debug-logging = false
-
   proxy {
-    # Hostname of the proxy. If undefined ("") proxy is not enabled.
+    # hostname of the proxy. If undefined ("") proxy is not enabled.
     host = ""
     port = 8000
   }
 
+  # default values for AWS configuration. If credentials and/or region are not specified when creating S3Client,
+  # these values will be used.
   aws {
     access-key-id = ""
     secret-access-key = ""
