+++ s3/src/main/scala/akka/stream/alpakka/s3/S3Exception.scala
@@ -3,7 +3,7 @@
  */
 package akka.stream.alpakka.s3
 
-import scala.xml.{Elem, NodeSeq, XML}
+import scala.xml.{Elem, XML}
 
 class S3Exception(val code: String, val message: String, val requestID: String, val hostId: String)
     extends RuntimeException(message) {
