+++ s3/src/main/scala/akka/stream/alpakka/s3/S3Settings.scala
@@ -11,13 +11,12 @@ final case class Proxy(host: String, port: Int)
 
 final class S3Settings(val bufferType: BufferType,
                        val diskBufferPath: String,
-                       val debugLogging: Boolean,
                        val proxy: Option[Proxy],
                        val awsCredentials: AWSCredentials,
                        val s3Region: String) {
 
   override def toString: String =
-    s"S3Settings($bufferType,$diskBufferPath,$debugLogging,$proxy,$awsCredentials,$s3Region)"
+    s"S3Settings($bufferType,$diskBufferPath,$proxy,$awsCredentials,$s3Region)"
 }
 
 sealed trait BufferType
@@ -44,7 +43,6 @@ object S3Settings {
       case _ => throw new IllegalArgumentException("Buffer type must be 'memory' or 'disk'")
     },
     diskBufferPath = config.getString("disk-buffer-path"),
-    debugLogging = config.getBoolean("debug-logging"),
     proxy = {
       if (config.getString("proxy.host") != "")
         Some(Proxy(config.getString("proxy.host"), config.getInt("proxy.port")))
